Feature/add srj (#205)
Relates to
https://www.notion.so/monsuivijustice/Ajouter-le-SRJ-l-appli-agents-97ef74365bf844de991992c6afca3b12

Changed schema to integrate SRJ and created City, Tj and Spip models.
Tj and Spip will have to be manually linked to MSJ organizations.
Added admin dashboard to administrate these three models.

<img width="1617" alt="Capture d’écran 2023-01-24 à 15 23 43"
src="https://user-images.githubusercontent.com/25039335/214319854-e7bf9736-6117-42f9-95bb-24f583ac2bec.png">

+++ app/controllers/admin/seeds_controller.rb
@@ -13,7 +13,11 @@ module Admin
 
     def reset_db
       sign_out(true_user)
+      sign_out(current_user)
       ActiveRecord::Base.connection.tables.each do |t|
+        # We don't want to delete the SRJ tables
+        next if %w[cities tjs spips commune structure type_structure ln_commune_structure].include?(t)
+
         conn = ActiveRecord::Base.connection
         conn.execute("TRUNCATE TABLE #{t} CASCADE;")
         conn.reset_pk_sequence!(t)
